Now, variable modifications are also working!
GetFixedPTM is converted to GetPTMs to select any PTMs either fixed or variable

Instead of checking existence of crosslinked database, now it checks a SETTINGS file in terms of database construction (of course whenever a program starts from scratch, it writes a setting file). If there is any change on settings, it creates xlinked database! 

Two columns for modifications for peptideA and peptideB are added to output file

Underscore for naming at method of getTheoreticalCXPeaks() at MatchAndScore

linkerposition information is fixed on FASTCPDBLoader